feat(AppPanel): display example app in a panel
The example app is now displayed in the middle of the screen. The
Livestax theme has been included so that the panel component can be
used.

A flexbox based layout is used to ensure that the app displays in the
middle of the screen. A `VerticalPanelBlock` component has been created.
This will host the panels that take up the same height as the app.

The app frame is 317px wide, the VerticalPanelBlock blocks flex around
it.

To ensure that styles can be written in less, `less-loader`,
`css-loader` and `style-loader` are all added as development
dependencies.