test(setup): allowing testing of jsx components with jsdom
The tests have been set up using `mocha` and `chai` for the expect
syntax.  Since jsx files can't be run natively in `mocha`, a compiler
is used.  This is a simple jsx transform using the `react-tools`
library.

To allow the testing for components being rendered, `jsdom` is used for
the DOM implementation. It has been locked to v3 since v4 only supports
io.js and not node.js