feat(AppBlocker): show loading state until app is ready
A "ready" signal is expected from an app before it can be displayed.
Since this requires changing the state of the React app, the Flux
pattern has been implemented using a Store and Actions on that store.

Since the "ready" signal comes through as a message on window, a
singleton `WindowListeners` module has been created to listen to
events on the window. It triggers a Flux action when one is received.
This keeps the functionality of this minimal and the actual logic can
happen in the store.

Currently there is no check for the type of message.

To allow the state in the store to be replaced or reset easily (either
doing a full application reload, or resetting the values for testing)
`reset` and `replaceState` functions have been added to the AppStore.