feat(AppBlocker): Allow a timeout state to be set
The `AppStore` now starts a timer when the `reset` method is
called (`reset` is called from the `AppStore` constructor).

If the timer completes before a `ready` message is received then
State.status` is set to `timeout`. When the `AppStore` enters this
state, the `AppBlocker` component is displayed with a "timeout" message.

If the `AppStore` receives a `ready` message before the timer completes
then the timer is cancelled.