feat(AppConfigurator): allow the app to be changed
Previously, the displayed app was always the example app. A new "App
Configuration" panel has been created, it contains a form that when
submitted, will update the details (mainly the url at the moment) of the
app.

Since the app is updated, it now uses the url returned from the app
store.

In most (all?) browsers, if the `src` attribute of an iframe is set to
the same value, it will not trigger a reload of the iframe. To fix this,
the iframe has been abstracted into an `AppIframe` component, this
component will set the src to "about:blank" briefly before setting it to
the actual state. This only happens when the state is not "ready". This
prevents the iframe from reloading unless triggered from an action.