feat(AppConfigurator): validate form fields
The AppConfigurator now validates the form fields when the form is
submitted. This is performed by a `validate` function on a new
`Validator` module. The `validate` function expects an immutable state
and a set of validation functions to perform, it returns a map
containing the keys and the errors returned from the validation
function.

Each validation function expects a field argument and returns either a
string if there is an error, or undefined if there is not an error.

The AppConfigurator sets the errors on its internal state if there are
any errors, these errors are then passed to the `Input` component and
displayed underneath the form field.