feat(MessageGenerator): add a message generator panel
A new panel has been added for the purpose of generating messages that
are sent to the app with `postMessage`.

`AppIframe` subscribes to change events from the `MessageStore`, and
when received it sends a message to the iframe using `postMessage`. If
`postMessage` is not defined on the `contentWindow` on the iframe (i.e.
if using jsdom in the test environment), then no message is sent. This
 function is stubbed in the tests.