chore(Gruntfile.js): production build task
Grunt has been added as a dependency to allow for a production build to
be created. A new webpack production configuration has been created, it
simply copies the webpack config and removes the entry points for the
webpack dev server, it also removes the plugins for hot-code reloading
and error checking.