feat(History): track history for events from panels
A new `History` component has been created, it retrieves its history
from the `HistoryStore` and triggers a user defined action when a
history item is clicked. The history can also be deleted.

The `HistoryStore` persists the changes in localStorage so that they are
available the next time the user visits the site.

`localStorage` is replaced with a simple JavaScript object in the test
environment.

The `HistoryStore` just has to listen to the events that it wishes to
generate history items for and store them against the relevant key. A
name field is currently required as this is what is used to display in
the list item.