feat(History): display timestamp for history items
History items are now stored with a unix timestamp representing the
`createdAt` datetime for the history item.

History items are now rendered in a table to be more consistent with the
Livestax theme. Each item including the timestamp.

Additionally History items for `Messages` have been updated to
incorporate the `value` of the `Message` in the generated History item
name.