feat(InputHorizontal): delegate additional props to input
To allow additional attributes to be set on the input HTML element the
`InputHorizontal` component uses destructuring assignment to strip out
the fields that are used directly by the component (error, name, label,
value, onChange) and pass the additional props to the input element.

To ensure that the destructuring assignment can be used in the tests,
the experimental flag has been enabled in the babel compiler when using
mocha.

Docs for destructuring assignment are available at:

  https://developer.mozilla.org/en-US/docs/Web/JavaScript/ \
  Reference/Operators/Destructuring_assignment