feat(SignedRequest): panel to input signed request data
A new panel has been created that allows the signed_request fields
secret_key, app_type, instance_id, timestamp, user_type, user_id
to be updated.

When an app is using the signed request, a hidden form is rendered
inline with the iframe, the target of the form is the name attribute of
the iframe. When the form submits, the app is loaded with the correct
url. This method allows the signed request to be POSTed into the app.

Since the signed request used in Livestax is an open standard, the
`simple-jwt` library is used to sign the post data with the secret key.

If the user type is set to guest then the user id field is disabled.
When the signed_request is POSTed, the user_id value is set to null, as
Livestax does not expect a user_id for a guest user.