feat(RadioHorizontal): component for displaying radio buttons
A new `RadioHorizontal` component has been created. This component has a
similar API to the checkbox, however since radio buttons are displayed
in groups, there is no label. Instead, `RadioHorizontal` should be
nested inside of another component. An `InputGroupHorizontal` component
allows for the radio buttons to be displayed in a horizontal form with a
label for the radio button group.