feat(KeyValueStore): flux store for handling store postMessages
Apps in the simulator can now register values against particular keys.
These events are sent via postMessage and are known as "store" events.
Since the word "store" is used in the Flux architecture, this particular
store has been named the KeyValueStore.

The KeyValueStore emits a change event when an app calls a `watch` or
`get` event and a value is set for that particular key. This in turn
emits a change that the `AppIframe` component is listening to. It then
transmits a message down to the app iframe via postMessage. Where
applicable, the namespace of the current app loaded into the iframe is
used.

If an app configuration event is received, the KeyValueStore will clear
the listeners that are set up, however it will keep the values. This
allows for "App A" to be loaded and set a value. Then for "App B" to be
loaded and retrieve the value, but if "App B" is loaded multiple times,
it does not create an additional listener per load.