feat(StorePanel): allow a value to be set in the Key-Value Store
The StorePanel contains a form that expects a key and value. When this
is submitted, it will set the value in the JavaScript listener store,
which will in turn trigger an event if any callbacks are set on that
key. The KeyValueStore listens from an event from the form and, if
appropriate, converts the string to JSON.