feat(history): condense history into a single panel
Previously the history for the app configuration panel and the message
generator were displayed in separate panels. This has been updated to
display all history items in one panel.

The `HistoryStore.getHistory` function has been updated to return a
single list of all history items, if no `key` parameter is provided.
Each item in the all history list is transformed to include a
`historyType` attribute. This attribute is utilised at various stages of
the history items life cycle.

When a request to delete a history item is made it is no longer
identified via its list index, instead a strict match done on the entire
object against all items in the store.

The History table display has been updated and includes icons for each
history type.