feat(StorePanel): display current store values
The JavaScript store has been modified to expose a function to get all
of the keys and values. Theses values can be retrieved from the
`KeyValueStore` and are returned as a JavaScript object. A new
`StoreTable` component has been created to display this data. If JSON
data is stored then it is stringified for display purposes.