feat(StorePanel): allow values to be unset from the store
The table rows in the StoreTable now trigger the `onDelete` callback
that is passed through from the StorePanel. When a row is deleted, it
unsets the value in the internal store, which triggers the callback in
the app. A change event is emitted to re-render the StoreTable without
the removed table row.