feat(SignedRequest): combine AppConfigurator with SignedRequest
The `SignedRequest` component has been merged into the `AppConfigurator`
component. The signed request section of the form is hidden by default
but can be toggled on/off with a checkbox. The validation rules for the
app configuration form are dependent upon whether the signed request
section is hidden.

This change has resulted in the removal of the `receiveSignedRequest`
action as the `receiveAppConfiguration` now includes a payload with the
signed request data.

If the app is configured to not use a signed request then the associated
history item is populated with blank signed request data. When this
history item is retrieved the signed request data is populated with
default data.