feat(history): add filters for history type
The history component now includes a filter menu which provides the user
with a method of filtering the history items by type.

The default filter is "All", which displays all history items of all
types.

The `HistoryStore.getHistory` function has been updated to always inject
a `historyType` property into history items. Previously this was only
done when "all" history items were requested; but it is now also done
when a specific history type is requested.

A new public function has been added to the `HistoryStore`, called
`getHistoryTypes` this returns a list of history types and is used when
constructing the filter menu.