feat(history): only show filters that contain items
Previously the history component displayed all available filters
regardless of whether items existed for that history type. This has been
updated so that only filters that are not empty are displayed.

Additionally when the last item within a history filter category is
deleted the active filter will be set to "All"