feat(logger): only log whitelisted types and subtypes
Previously, all `postMessages` types were added to the logger, however,
there are some that would be best hidden. This is due to the inability
for app developers to utilize these, such as `scroll`.

For added protection, there is a check on the subtype to ensure only
desired logs are added. However, some types allow for "wildcard"
string as their subtype, therefore this is denoted by a "*".