feat(LoggerStore): set direction on postMessage received
When a postMessage action is received, the direction of the message
added to the store is set to "from" to indicate that the message has
come from the app. It is safe to assume that anything that is not "from"
is considered a "to" message as there are only two directions in which a
message can be sent.