feat(Logger): format logger list
The logger list is now presented as a table. With each row broken down
into the following format:

| type | subtype | payload |

Previously only the top level of the `LoggerStores` state was an
Immutable object (its nested properties weren't Immutable). The `Logger`
component had a requirement to apply transformations to the `payload`
object; it was found to be more convenient to apply these
transformations if the source was an Immutable structure. Based on this
the decision was made to make the `LoggerStores` state fully Immutable.
This was achieved by using `fromJS` which does a deep transformation.