feat(LoggerStore): listen for message generator actions
When an action is triggered from the message generator, the LoggerStore
now listens to the event and appends a log for the event. Since the
payload that the app is expecting does not exactly match the event
thrown by the MessageGenerator it needs to be modified into an app
payload. This was previously done in the AppIframe component, however
since this is now also used in the LoggerStore, a Projections file has
been created to perform this transformation.

Due to the way the generator Projection returns the data, with the top
level type, it means that the store payload did not match correctly. A
projection has also been created for the store payload to ensure the
data is consistent for sending through postMessage.