feat(Logger): auto-scroll logs if scrolled to bottom
The Logger component checks the scroll position of its internal panel
before the component updates. It sets a flag `shouldScroll` based on if
the Logger has been scrolled to the bottom on the componentWillUpdate
lifecycle function. On componentDidUpdate it will scroll the Logger to
the bottom if the `shouldScroll` flag is set.