refactor(History): extract filters into a component
Previously the responsibility for rendering a list of filters was held
by the `History` component. This functionality has been extracted into
a new reusable component called `FilterList` which is used by the
`History` component.

The `History` component is responsible for telling the `Filter`
component which filters should be displayed, and which filter is
currently active. The `FilterList` is purely concerning with rendering
the filters list.