fix(FilterList): pass all required props to the component in test
Previously one of the `FilterList` tests didn't supply all the required
props when instantiating the component, as a result a React warning was
being raised.

The test has been updated to ensure all required props are
passed to the component and as a result the warning is no longer raised.