fix(FilterList): add unique key to each filter item
Previously when rendering the filter list the `FilterList` component
assigned a `key` to each filter element based on value of the filter
items `data` attribute, which was of type `String`.
React did not deem these values to be unique and raised an `Each child
in an array should have a unique "key" prop.` warning.

The key is now based off the current `index` of the iteration which
ensures that a warning is not raised.