feat(collapsible panel): add toolbar to allow for actions
Previously, there was no method of displaying a toolbar in a panel,
however, the `PanelToolbar` component allows for a toolbar to be
rendered in a fixed position above the the content.

Previously, the Clear button in the Logger Panel was a `<button>` with
different styles which made it inconsistent with the filter toggles.
The Clear button now uses a `<span>` tag for consistency.