feat(FlashMessage): trigger flash callback when flash interaction occurs
The `FlashMessageStore` registers an interest in `FLASH_INTERACTION`
actions. When such an action is received the store sets the
`interaction` property of it's state to reflect the interaction type.

The `AppIframe` component listens for `change` events from the
`FlashMessageStore`. When a `change` event is received the component
checks is an interaction is present and if so sends a post message to
the app. This will ensure that any flash message callbacks are triggered
within the app.