fix(FlashMessage): clear flash message when new app config received
Previously when a new app configuration was loaded the `FlashMessage`
panel wasn't cleared. This has been fixed by purging the
`FlashMessageStore` state whenever a `RECEIVE_APP_CONFIGURATION` action
is received.

Additionally it was necessary to modify the structure of the
`FlashMessage` component's internal state so that it was an object (as
opposed to a simple string). This was necessary for `null` changes to be
registered when using `replaceState`.