style(FlashMessage): add colorized flash header to denote flash type
The `FlashMessage` now incorporates a flash header element with a
background-color that is based on the flash message type.

Additionally the flash message buttons have been styled so that the
focused button does not look visually distinct from other buttons. As
previously once a button was clicked it appeared to remain selected.