feat(toolbar): move toolbar to own panel
Previously, the toolbar was rendered within the `panel-body`, however,
it required the toolbar to be in a fixed position which lost required
inheritance such as width.

Now, it is rendered within its own panel inside the collapsible panel
and sits above the `panel-body` so the body content falls under it when
scrolling. This still allows for the panel to be hidden.

As there is a second `panel-collapse`, the target for scrolling to the
bottom for the logger needed to be updated. This has now been named
`panel-body-container` to remain consistent with
`panel-toolbar-container`.