feat(FlashMessage): disable buttons with no callbacks
When a postMessage for the flash message is received, if it contains a
callbacks array, any buttons not present in the array will be disabled.
If a callbacks property is not present in the postMessage payload (due
to using an older version of the API) then it is defaulted to show all 3
callbacks.