feat(FlashMessage): allow flash message to be cleared
In a previous commit, the FlashMessage interaction buttons were toggled
between enabled and disabled depending on if the app developer has set a
callback.

There is a case that the developer could not have set up any callbacks,
resulting in all of the buttons being disabled. Since the flash message
is cleared on interaction, and there is no way to interact, the flash
message will remain in the panel until another flash message arrives. A
clear button is now present on the flash message for this case.