fix(History): fix history test for deleting items
In commit 25c751b58f5b7776130f039b74c27d9ee9275fa8 a key was added to
the `tr` for each history row with the value `i`.

Since this numeric key was being used, each simulated click event was
triggering a re-render, for 3 table rows it would look like:

| key | row   |
|-----|-------|
| 0   | foo   |
| 1   | bar   |
| 2   | baz   |

After the first delete (index 0) it would be:

| key | row   |
|-----|-------|
| 0   | bar   |
| 1   | baz   |

After the second delete (index 1) it would be:

| key | row   |
|-----|-------|
| 0   | bar   |

When attempting a 3rd delete with index 2, React would complain that the
DOM node is not mounted, this is correct as the key does not match the
index.

The fix is to reverse the order of the click events, however explicitly
specifying the index was preferred instead of calling `reverse` on the
array to be more explicit.