feat(AppConfigurator): allow non-admin user type
Previously the app configurator allowed for the following cases:

| Type  | user_id | is_admin | is_guest |
|-------|---------|----------|----------|
| Guest | null    | false    | true     |
| Admin | UUID    | true     | false    |

There was not a case that allowed for a normal user - that is a user who
is logged into the system but not an admin of the space. This has been
added to the form, but is considered checked when both `is_admin` and
`is_guest` are false. This ensures that there isn't an additional
`is_user` property passed through in the payload in the simulator that
does not exist in the platform. The new values are:

| Type  | user_id | is_admin | is_guest |
|-------|---------|----------|----------|
| Guest | null    | false    | true     |
| User  | UUID    | false    | false    |
| Admin | UUID    | true     | false    |