feat(MenuPanel): panel to display menu items
The MenuPanel currently displays the "Reload" menu item. When "Reload"
is clicked, the current app is reloaded with its original app
configuration. The `receiveAppConfiguration` listener on the `AppStore`
has been modified slightly to use the existing configuration if no
`data` is passed with the action. Using the existing action instead of
creating a `reloadApp` action is used because other panels (e.g. the
`FlashMessage` modify their state based on the `receiveAppConfiguration`
action.

The "Reload" item is a special case, it is not added from inside the
app, but instead created by the platform as the functionality is not
directly related to the app.