feat(MenuPanel): display menu items from the app
The `MenuPanel` now gets additional menu items from the `MenuStore`.
These are added as they come from the app. An OrderedSet is used to
preserve order but discard duplicate menu items. The platform filters
these out anyway, however it is an additional guard against it.

If no icon is set then a default of "circle-o" is used. Currently
clicking on a custom menu item just triggers a noop.