feat(MenuPanel): trigger custom menu events
When a custom menu item is clicked, an event is sent to the app. The
event contains the name of the menu item clicked, this is used by the
app to trigger the correct callback.

Since menu events are transient, the name of the menu item is sent to
the `AppIframe` on the change event. If there is no name specified, the
event is ignored, otherwise it sents the postMessage to the app.