feat(MenuPanel): reset menu items when app internally navigates
The `MenuStore` now registers an interest in `ready` post messages. When
such a message is received the store resets the menu `items` to be an
empty set. This ensures that if an app internally navigates to a new
page then the menu items will be cleared. This behaviour mimics the
behaviour of Livestax.