feat(History): add clear button to panel toobar
The `History` component now includes a `clear` button in the
`PanelToolbar`. When clicked this will trigger an action that clears the
history items stored in `localStorage`.

In the test environment, `localStorage` is a test double. This double
has been updated to include a simple `clear` function, that mimics the
behaviour of its real counterpart.

The `History` component tests for filter functionality, have been
updated as the finders within these tests were too general and were
being effected by the presence of the new `clear` label.