feat(Dialog): add dialog panel
The `DialogPanel` component listens to change events from the
`DialogStore`. When a change event is broadcast the dialog data is
fetched from the store and the text content is rendered.

The `DialogStore` subscribes to incoming post messages and when a dialog
message is received, it updates its state accordingly.

If no dialog message has been received then the `DialogPanel` will
display an `EmptyPanel`, which renders a simple help message.

The example app has been updated to include a button, that when clicked
will call `Livestax.dialog.show`