feat(Dialog): trigger dialog callbacks when dialog buttons are clicked
The `DialogPanel` now triggers a Dialog action, when a user clicks a
dialog button. This produces a change event to be broadcast from the
`DialogStore`.

The `AppIframe` component subscribes to `DialogStore` change events.
When such an event is received the `AppIframe` component sends a post
message down to the app, which will trigger the appropriate callback.