feat(Dialog): clear dialog when "clear" is clicked in panel toolbar
A `PanelToolbar` has been added to the `DialogPanel`. This toolbar
contains a `clear` button. When clicked this triggers a dialog action,
that the `DialogStore` subscribes to. The store then triggers a change
action which will result in a post message being passed down to the app.

The payload for the message has a `title` attribute with a `cancel`
value.  This will ensure that any callbacks registered with buttons that
have a "cancel" title will be triggered.

It is necessary to send a post message to the app as this ensures that
Livestax will purge the registered callbacks, allowing new dialogs to be
shown.