feat(banner): add Livestax branding banner
Livestax products are being branded with the same branding included in
the marketing material. This banner addition brings the simulator in
line with the branding guidelines.

The addition of the banner causes the panel toolbars to potentially drop
below the recommended height of `28px` so a `min-height: 28px` has also
been added to ensure the correct look.