feat(authenticate): display authentication provider name in panel
A new `AuthenticateStore` has been created that registers an interest in
`authenticate` post messages. On receiving such an action the store
updates its state and broadcasts a `change` event.

A new `AuthenticatePanel` has been added to the simulator. This
component subscribes to `change` events from the `AuthenticateStore`. On
receiving such an event the panel renders the authentication providers'
name.

The `AuthenticateStore` also registers an interest in
`RECEIVE_APP_CONFIGURATION` actions. On receiving such an action the
store state is cleared, this ensures the panel will be cleared of any
previous authenticate request data.

The `AuthenticatePanel` renders an `EmptyPanel` with help message, if
no `authRequest` is present.