feat(authenticate): open authenticate window
On receiving an `authRequest`, the `AuthenticatePanel` now displays an
`Authenticate` button. When this button is clicked an
`OPEN_AUTHENTICATE_WINDOW` action is triggered.

The `AuthenticateStore` registers an interest in
`OPEN_AUTHENTICATE_WINDOW` actions, and on receiving such an action it
triggers an `openWindow` `AuthenticateIncident` which will open the
authenticate url in a new window.