feat(authenticate): handle authenticate response
The `AuthenticateIncident.openWindow` function now polls the popup
window every second with a `livestax:ping` post message. This will
occur until `Livestax.authenticate.respond` has been called, at which
point an `authenticate.response` post message is sent back from the
Livestax JavaScript API.

The `AuthenticateStore` registers an interest in `authenticate.response`
post messages, and on receiving such an action raises a `responded`
`AuthenticateIncident`, which stops the `livestax:ping` polling.

The `AppIframe` component now listens for `AuthenticateStore` change
events, and when such an event is associated with a
`authenticate.response` message. An `authenticate.respond` post message
is sent down to the app, which ensures that the callback registered with
`authenticate.start` is triggered.