fix(iframe): hide iframe until ready state
Previously, the App iframe was displayed when loading the app, however,
if the Livestax JS was not included in the app, the web app would still
display until the app would timeout.

Now, until the app is ready, the iframe has a `hidden` class so it won't
display. There was an issue where the iframe would reload when not
required, this has been avoided by only reloading the iframe when the
status changes from `loading` to `ready` and not just on status change.