feat(flash): remove flash message when `clear` message is received
Version 0.4.0 of the Livestax JavaScript API, provides an interface for
requesting that a flash message should be cleared. The current commit
ensures that the Simulator honors such a request and does remove the
flash message.

The example app has been updated to depend upon version 0.4.0 of the
Livestax API. This app has been modified to include features that
demonstrate the new `clear` functionality.